rust bindings: Updated the API
We now expose `InferenceContext::create_context(...)`
to create a new inference context for inference.
