+++ rust/src/lib.rs
@@ -1,11 +1,11 @@
-//! Bindings for exposing the functionality of librunecoral.a
+//! Bindings for exposing the functionality of `librunecoral`
 //!
 //! # Example
 //!
 //! ```rust,no_run
 //! # fn load_model() -> &'static [u8] { todo!() }
 //! # fn main() -> Result<(), Box<dyn std::error::Error>> {
-//! use hotg_runecoral::{Tensor, TensorMut, create_inference_context};
+//! use hotg_runecoral::{Tensor, TensorMut, InferenceContext};
 //!
 //! // load the model
 //! let model: &[u8] = load_model();
@@ -19,7 +19,7 @@
 //! let output_tensor = TensorMut::from_slice(&mut output, &[1]);
 //!
 //! // load our inference backend
-//! let mut ctx = create_inference_context(
+//! let mut ctx = InferenceContext::create_context(
 //!     "application/tflite-context",
 //!     model,
 //!     &[input_tensor.descriptor()],
@@ -37,12 +37,10 @@
 
 mod context;
 pub mod ffi;
-mod rune_coral;
 mod tensors;
 
 pub use crate::{
-    context::InferenceContext,
-    rune_coral::{LoadError, create_inference_context},
+    context::{InferenceContext, LoadError},
     tensors::{ElementType, Tensor, TensorDescriptor, TensorElement, TensorMut},
 };
